Moved TaskBoardSampleData into assets
